remove static IPs from seed lists
