+++ mm2src/mm2_net/src/ip_addr.rs
@@ -1,5 +1,6 @@
 use crate::transport::slurp_url;
-use common::log;
+use common::{cfg_native, log};
+use derive_more::Display;
 use gstuff::try_s;
 use gstuff::{ERR, ERRL};
 use mm2_core::mm_ctx::MmArc;
@@ -9,6 +10,11 @@ use std::fs;
 use std::io::Read;
 use std::net::{IpAddr, Ipv4Addr};
 use std::path::Path;
+
+cfg_native! {
+    use std::net::ToSocketAddrs;
+    use mm2_err_handle::prelude::MmError;
+}
 
 const IP_PROVIDERS: [&str; 2] = ["http://checkip.amazonaws.com/", "http://api.ipify.org"];
 
@@ -168,4 +174,50 @@ pub async fn myipaddr(ctx: MmArc) -> Result<IpAddr, String> {
         try_s!(detect_myipaddr(ctx).await)
     };
     Ok(myipaddr)
+}
+
+#[derive(Debug, Display)]
+pub enum ParseAddressError {
+    #[display(fmt = "Address/Seed {} resolved to IPv6 which is not supported", _0)]
+    UnsupportedIPv6Address(String),
+    #[display(fmt = "Address/Seed {} to_socket_addrs empty iter", _0)]
+    EmptyIterator(String),
+    #[display(fmt = "Couldn't resolve '{}' Address/Seed: {}", _0, _1)]
+    UnresolvedAddress(String, String),
+}
+
+#[cfg(not(target_arch = "wasm32"))]
+pub fn addr_to_ipv4_string(address: &str) -> Result<String, MmError<ParseAddressError>> {
+    // Remove "https:// or http://" etc.. from address str
+    let formated_address = address.split("://").last().unwrap_or(address);
+
+    let address_with_port = format!(
+        "{formated_address}{}",
+        if formated_address.contains(':') { "" } else { ":0" }
+    );
+
+    match address_with_port.as_str().to_socket_addrs() {
+        Ok(mut iter) => match iter.next() {
+            Some(addr) => {
+                if addr.is_ipv4() {
+                    Ok(addr.ip().to_string())
+                } else {
+                    log::warn!(
+                        "Address/Seed {} resolved to IPv6 {} which is not supported",
+                        address,
+                        addr
+                    );
+                    MmError::err(ParseAddressError::UnsupportedIPv6Address(address.into()))
+                }
+            },
+            None => {
+                log::warn!("Address/Seed {} to_socket_addrs empty iter", address);
+                MmError::err(ParseAddressError::EmptyIterator(address.into()))
+            },
+        },
+        Err(e) => {
+            log::error!("Couldn't resolve '{}' seed: {}", address, e);
+            MmError::err(ParseAddressError::UnresolvedAddress(address.into(), e.to_string()))
+        },
+    }
 }
