+++ mm2src/mm2_p2p/Cargo.toml
@@ -22,6 +22,7 @@ lazy_static = "1.4"
 log = "0.4"
 mm2_core = { path = "../mm2_core" }
 mm2_event_stream = { path = "../mm2_event_stream" }
+mm2_net = { path = "../mm2_net" }
 mm2_number = { path = "../mm2_number", optional = true }
 parking_lot = { version = "0.12.0", features = ["nightly"] }
 rand = { version = "0.7", default-features = false, features = ["wasm-bindgen"] }
